episode 7 movement ancora uno schifo, smartenemy, casino da int a float

+++ src/game/BasicEnemy.java
@@ -27,11 +27,11 @@ public class BasicEnemy extends GameObject {
 
     public void render(Graphics g) {
         g.setColor(Color.red);
-        g.fillRect(x, y, 16, 16);
+        g.fillRect((int) x, (int) y, 16, 16);
     }
 
     public Rectangle getBounds() {
-        return new Rectangle(x, y, 16, 16);
+        return new Rectangle((int)x ,(int) y, 16, 16);
     }
     
 }
