+++ src/game/GameObject.java
@@ -4,11 +4,11 @@ import java.awt.Graphics;
 import java.awt.Rectangle;
 
 public abstract class GameObject {
-    protected int x, y;
-    protected int velX, velY;
+    protected float x, y;
+    protected float velX, velY;
     protected ID id;
 
-    public GameObject(int x, int y, ID id) {
+    public GameObject(float x, float y, ID id) {
         this.x = x;
         this.y = y;
         this.id = id;
@@ -19,36 +19,36 @@ public abstract class GameObject {
     public abstract Rectangle getBounds();
 
     //setters
-    public void setX(int x) {
+    public void setX(float x) {
         this.x =  x;
     }
-    public void setY(int y) {
+    public void setY(float y) {
         this.y = y;
     }
     public void setId(ID id) {
         this.id = id;
     }
-    public void setVelX(int velX) {
+    public void setVelX(float velX) {
         this.velX = velX;
     }   
-    public void setVelY(int velY) {
+    public void setVelY(float velY) {
         this.velY = velY;
     }
 
     //getters
-    public int getX() {
+    public float getX() {
         return x;
     }
-    public int getY() {
+    public float getY() {
         return y;
     }
     public ID getId() {
         return id;
     }
-    public int getVelX() {
+    public float getVelX() {
         return velX;
     }
-    public int getVelY() {
+    public float getVelY() {
         return velY;
     }
